feat(vulnerable-groups): refactored related code
